+++ backend/apps/inventory/api.py
@@ -7,7 +7,7 @@ from .models import Inventory, Issue, Receipt
 class StockSerializer(serializers.ModelSerializer):
     product_code = serializers.CharField(source="product.code", read_only=True)
     product_name = serializers.CharField(source="product.name", read_only=True)
-    available_quantity = serializers.IntegerField(read_only=True)
+    available_quantity = serializers.DecimalField(max_digits=18, decimal_places=6, read_only=True)
 
     class Meta:
         model = Inventory
